Validate context for email address in SendEmailHandler
The /api/v1/mailers/send-email API now validates context separately fo
each email address.

Previously the handler validated context globally, which was not
intentional. It could have prevented a notification with the same context
to be sent to different users. This change unifies the behavior with jobs
- they checked the email-context pair since the beginning.

To correctly test the behavior, Mockery library was added as a dependency.

remp/crm#2226