Pass email params to the recipient variables for non-batch sending
Originally, there was a discrepancy in the X-Mailer-Template-Params
message header:

- Batch sending used the structure with email addresses:

  {
      "user1@example.com": { /* variables */ },
      "user2@example.com": { /* variables */ }
  }

- Non-batch sending used the more direct variant

  {
      /* variables */
  }

We fixed this discrepancy in a hotfix in e1a02216. On 10-June-2022
Mailgun started to return HTTP 400 for the recipient variables
structured as in the example 2 and the change was necessary.

This commit brings the temporarily removed params back to the header.

We also changed the way how message ID is generated to use more
faster crc32b hash instead of the fully random string.

remp/remp#1146