Fix UnreadArticlesResolver which crashed job processing
The processing crashed in case of an uncaught (invalid URL) exception.
Such scenario could happen if an article was unpublished, but remained
in the stats and was selected for personalized newsletter.

Mailer wouldn't be able to fetch meta for article and some content
resolvers throw InvalidUrlException if they cannot access the article.
The exception itself is correct, because we should differentiate
if the URL is invalid or if the URL is valid, but without processable
metadata.

remp/remp#1017