Fixes removal of template (Twig) variables in Mailer
When editing HTML email template and switching from HTML to WYSIWYG editor and back, editor broke Twig variables placed in other than text contexts. 

For example, have this code:

```
<a href="#" style="color:{{ include("c1") }}">
```

This is not a valid HTML code, but it is a valid code in our context - Twig variable in the style attribute will be replaced with a return value of the Twig function. However, WYSIWYG editor previously broke the syntax of this code - the problem were double quotes inside double quotes.

Current solution replaces all Twig variables in non-text contexts with text replacements that are kept untouched by the WYSIWYG editor. These replacements are switched back to Twig variables when user returns from WYSIWYG editor to HTML editor.

This solution might not work in all cases, but it works in most cases we have encountered.

remp/remp#719