Changed Sender condition including HTML/text version of email
Before the change, the inclusion was based on the presence
of HTML/text content in mail template. This wasn't always OK
and could cause issues if the template was empty and only
layout was used to send an email.

The content was there, but not in the tested mail template.

remp/remp#1392